+++ ereuse_devicehub/resources/documents/device_row.py
@@ -235,6 +235,7 @@ class DeviceRow(OrderedDict):
             self['Test {} {} Power on (hours used)'.format(ctype, i)] = ''
             self['Test {} {} Lifetime remaining (percentage)'.format(
                 ctype, i)] = ''
+            self['Test {} {} Power on hours'.format(ctype, i)] = ''
             return
 
         snapshot = get_action(component, 'Snapshot')
@@ -332,6 +333,7 @@ class DeviceRow(OrderedDict):
             self['Test {} {} Power on (hours used)'.format(ctype, i)] = ''
             self['Test {} {} Lifetime remaining (percentage)'.format(
                 ctype, i)] = ''
+            self['Test {} {} Power on hours'.format(ctype, i)] = ''
             return
 
         self['Test {} {} Software'.format(ctype, i)] = software
@@ -342,6 +344,8 @@ class DeviceRow(OrderedDict):
             test_storage.power_cycle_count)
         self['Test {} {} Lifetime remaining (percentage)'.format(ctype, i)] = none2str(
             test_storage.lifetime)
+        self['Test {} {} Power on hours'.format(ctype, i)] = none2str(
+            test_storage.power_on_hours)
 
     def get_graphic_card(self, ctype, i, component):
         """Particular fields for component GraphicCard."""