+++ CHANGELOG.md
@@ -17,6 +17,7 @@ ml).
 - [added] #167 new actions of status devices: use, recycling, refurbish and management.
 - [changes] #177 new structure of trade.
 - [bugfix] #184 clean nested of schemas of lot
+- [added] #182 adding power on hours
 
 ## [1.0.9-beta]
 - [added] #159 external document as proof of erase of disk